Refactor clean and improve API.
